made updates and changes
