+++ src/components/ViewModal/ViewModal.module.css
@@ -1,27 +1,10 @@
-.parent {
-  background: gray;
-  height: 100vh;
-  width: 100%;
-  position: relative;
-}
-
-.active {
-  border: #CBCBCB solid 1px;
-  border-radius: 25px;
-  width: 50%;
-  height: 62vh;
-  position: absolute;
-  top: 20%;
-  left: 25%;
-}
-
 .bodyLeft {
   position: absolute;
   display: flex;
   flex-direction: column;
   max-height: 35vh;
   width: 50%;
-  top: 43%;
+  top: 48%;
   left: 5%;
 }
 
@@ -29,7 +12,7 @@
   position: absolute;
   max-height: 35vh;
   width: 35%;
-  top: 43%;
+  top: 48%;
   left: 62%;
   display: flex;
   flex-direction: column;
@@ -39,6 +22,11 @@
   top: 2%;
   left: 96%;
   position: absolute;
+}
+
+.height {
+  height: 80vh !important;
+  max-height: 80vh !important;
 }
 
 .iconText {
@@ -51,8 +39,8 @@
 
 .joinButton {
   position:absolute;
-  top: 112%;
-  left: 27%;
+  top: 140%;
+  left: 35%;
   border: #CBCBCB solid 1px;
   width: 100px;
   height: 40px;
@@ -67,7 +55,7 @@
   flex-direction: column;
   row-gap: 2px;
   margin-top: 0px;
-  margin-left: -22px;
+  margin-left: 0px;
 }
 
 .listText {
@@ -87,7 +75,7 @@
   top: 5%;
   left: 10%;
   width: 80%;
-  height: 15vh;
+  height: 25vh;
   object-fit:cover;
   overflow: hidden;
 }
@@ -111,8 +99,13 @@
 
 .title {
   position: absolute;
-  top: 30%;
+  top: 40%;
   font-size: 30px;
   text-align: center;
   width: 100%;
+}
+
+.width {
+  width: 60% !important;
+  max-width: 60% !important; 
 }
